+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug with maxOverZoom
 - Fixed bug with refreshing task list
 - Fixed server rendering (notification count, footer slot)
+- Fixed registrations with an invitation on sites without registrations enabled (MAD-1388)
 
 ### Changed
 - Add short descriptions for admin Search & Metadata actions + correct site config. page titles (MAD-1466)